Changes in data structure (hopefully the last!)
DUALC now uses firebase push() function to generate an unique ID for
every item. An internal "id" field was removed, for it's not necessary.
Classes and etc were adapted to work with this. There's still very ugly
things in the runtime data structures, but at least the JSON should
remain stable (and because of this, every version from now on should be
able to read the online data.)